+++ Arkanoid/Application/Utils/Game/GameInfo.cs
@@ -4,7 +4,7 @@ namespace Arkanoid.Application.Utils.Game
 {
     public struct GameInfo
     {
-        public float ElapsedFrameTime { get; set; }
+        public double ElapsedFrameTime { get; set; }
 
         public float ComputedSpeed { get; set; }
